Implemented task 1.
The DiceManager hold the number of rolls in the current turn and the
number of times a player has to roll. It also keeps track of the total.
To implement task 1, Die only calls startMoving once the roll number
equals the number of dice. Else, it propmts the player (or computer)
to roll again, whic then DiceManager registers the roll.